Logic for resetting password
