+++ backend/utils/generate_password.py
@@ -1,0 +1,4 @@
+import uuid
+
+def generate_password():
+    return uuid.uuid4().hex[:10]
